requirements.txt
sqlalchemy>=1.4.15
flask>=2.0.0
Jinja2>=3.0.0
werkzeug>=2.0.0
Pillow>=8.2.0
numpy~=1.20.3
tensorflow~=2.5.0